♻️refactor: common reminder 구조를 정리하고 전달 모델을 reminder 도메인으로 이동
