--- src/main/java/com/project/backend/domain/common/plan/factory/PlanEventFactory.java
@@ -1,75 +1,0 @@
-package com.project.backend.domain.common.plan.factory;
-
-import com.project.backend.domain.common.reminder.dto.PlanChanged;
-import com.project.backend.domain.common.reminder.dto.RecurrenceExceptionChanged;
-import com.project.backend.domain.common.reminder.dto.ReminderDeleted;
-import com.project.backend.domain.reminder.enums.ChangeType;
-import com.project.backend.domain.reminder.enums.DeletedType;
-import com.project.backend.domain.reminder.enums.ExceptionChangeType;
-import com.project.backend.domain.reminder.enums.TargetType;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@NoArgsConstructor(access = lombok.AccessLevel.PRIVATE)
-public class PlanEventFactory {
-
-    public static PlanChanged toPlanChanged(
-            Long targetId,
-            TargetType targetType,
-            Long memberId,
-            String title,
-            Boolean isRecurring,
-            LocalDateTime occurrenceTime,
-            ChangeType changeType) {
-        return PlanChanged.builder()
-                .targetId(targetId)
-                .targetType(targetType)
-                .memberId(memberId)
-                .title(title)
-                .isRecurring(isRecurring)
-                .occurrenceTime(occurrenceTime)
-                .changeType(changeType)
-                .build();
-    }
-
-    public static RecurrenceExceptionChanged toRecurrenceExceptionChanged (
-            Long exceptionId,
-            Long eventId,
-            TargetType targetType,
-            Long memberId,
-            String title,
-            Boolean isRecurring,
-            LocalDateTime occurrenceTime,
-            ExceptionChangeType changeType
-    ) {
-        return RecurrenceExceptionChanged.builder()
-                .exceptionId(exceptionId)
-                .eventId(eventId)
-                .targetType(targetType)
-                .memberId(memberId)
-                .title(title)
-                .isrRecurring(isRecurring)
-                .occurrenceTime(occurrenceTime)
-                .changeType(changeType)
-                .build();
-    }
-
-    public static ReminderDeleted toReminderDeleted(
-            Long exceptionId,
-            Long memberId,
-            LocalDateTime occurrenceTime,
-            Long targetId,
-            TargetType targetType,
-            DeletedType deletedType
-    ) {
-        return ReminderDeleted.builder()
-                .exceptionId(exceptionId)
-                .memberId(memberId)
-                .occurrenceTime(occurrenceTime)
-                .targetId(targetId)
-                .targetType(targetType)
-                .deletedType(deletedType)
-                .build();
-    }
-}
