+++ src/main/java/com/project/backend/domain/event/converter/RecurrenceGroupSpec.java
@@ -1,7 +1,6 @@
 package com.project.backend.domain.event.converter;
 
 import com.project.backend.domain.event.enums.MonthlyType;
-import com.project.backend.domain.common.plan.enums.MonthlyWeekdayRule;
 import com.project.backend.domain.event.enums.RecurrenceEndType;
 import com.project.backend.domain.event.enums.RecurrenceFrequency;
 import lombok.Builder;
