+++ src/main/java/com/project/backend/domain/nlp/dto/request/NlpReqDTO.java
@@ -1,6 +1,6 @@
 package com.project.backend.domain.nlp.dto.request;
 
-import com.project.backend.domain.common.plan.enums.MonthlyWeekdayRule;
+import com.project.backend.domain.common.recurrence.enums.MonthlyWeekdayRule;
 import com.project.backend.domain.event.enums.EventColor;
 import com.project.backend.domain.event.enums.MonthlyType;
 import com.project.backend.domain.event.enums.RecurrenceEndType;
