+++ src/main/java/com/project/backend/domain/todo/converter/TodoConverter.java
@@ -1,6 +1,6 @@
 package com.project.backend.domain.todo.converter;
 
-import com.project.backend.domain.common.plan.enums.MonthlyWeekdayRule;
+import com.project.backend.domain.common.recurrence.enums.MonthlyWeekdayRule;
 import com.project.backend.domain.event.exception.RecurrenceGroupErrorCode;
 import com.project.backend.domain.event.exception.RecurrenceGroupException;
 import com.project.backend.domain.member.entity.Member;
