+++ src/main/java/com/project/backend/domain/todo/dto/request/TodoReqDTO.java
@@ -1,7 +1,7 @@
 package com.project.backend.domain.todo.dto.request;
 
 import com.project.backend.domain.event.enums.MonthlyType;
-import com.project.backend.domain.common.plan.enums.MonthlyWeekdayRule;
+import com.project.backend.domain.common.recurrence.enums.MonthlyWeekdayRule;
 import com.project.backend.domain.event.enums.RecurrenceEndType;
 import com.project.backend.domain.event.enums.RecurrenceFrequency;
 import com.project.backend.domain.todo.enums.Priority;
